LayoutTests/fast/dom/Window/querySelectorAll-with-pseudo-elements.html

<!DOCTYPE html>
<html>
<head>
<title>Test for Bugzilla Bug 83446 - Broken handling for pseudo-elements in Selectors API </title>
<script src="../../js/resources/js-test-pre.js"></script>
<script>
function test() {
description("The test verifies that the querySelectorAll() API does not return any matched elements when querying for pseudo-element selectors.");

shouldBe('document.querySelectorAll("div::after").length', '0');
shouldBe('document.querySelectorAll("html::before").length', '0');

isSuccessfullyParsed();
}
</script>

<style type="text/css">
div:after {
    content:"";
}
</style>

</head>
<body onLoad="test()">
<p>Test for Bugzilla <a href="https://bugs.webkit.org/show_bug.cgi?id=83446">Bug 83446</a>: Broken handling for pseudo-elements in Selectors API.</p>
<div> </div>

</body>
</html>
