third_party/WebKit/LayoutTests/fast/dom/Window/querySelectorAll-with-pseudo-elements-expected.txt

The test verifies that the querySelectorAll() API does not return any matched elements when querying for pseudo-element selectors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS document.querySelectorAll("div::after").length is 0
PASS document.querySelectorAll("html::before").length is 0
PASS successfullyParsed is true

TEST COMPLETE


Test for Bugzilla Bug 83446: Broken handling for pseudo-elements in Selectors API.
